move wait_for_condition to tests.core
